Refactor MetaDataChangedListener
