+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/listener/GovernanceListenerManager.java
@@ -17,7 +17,6 @@
 
 package org.apache.shardingsphere.governance.core.registry.listener;
 
-import org.apache.shardingsphere.governance.core.registry.listener.metadata.MetaDataListener;
 import org.apache.shardingsphere.governance.repository.api.RegistryRepository;
 import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
@@ -37,14 +36,11 @@ public final class GovernanceListenerManager {
     
     private final Collection<String> schemaNames;
     
-    private final MetaDataListener metaDataListener;
-    
     private final Collection<GovernanceListenerFactory> governanceListenerFactories;
     
     public GovernanceListenerManager(final RegistryRepository registryRepository, final Collection<String> schemaNames) {
         this.registryRepository = registryRepository;
         this.schemaNames = schemaNames;
-        metaDataListener = new MetaDataListener(registryRepository, schemaNames);
         governanceListenerFactories = ShardingSphereServiceLoader.getSingletonServiceInstances(GovernanceListenerFactory.class);
     }
     
@@ -52,7 +48,6 @@ public final class GovernanceListenerManager {
      * Initialize all state changed listeners.
      */
     public void initListeners() {
-        metaDataListener.watch();
         for (GovernanceListenerFactory each : governanceListenerFactories) {
             each.create(registryRepository, schemaNames).watch(each.getWatchTypes().toArray(new Type[0]));
         }
